HW-5 Remove viewModel in BaseActivity
